마이페이지 해당 유저의 nickname, email, interest 값 프론트에 긁어오는데 성공.

+++ routes/mypage.router.js
@@ -5,59 +5,58 @@ const { Users, Posts, Comments } = require('../models');
 const bcrypt = require('bcrypt');
 const router = express.Router();
 
-router.get('/user/mypage/:userId', verifyAccessToken, async (req, res) => {
-  const userData = res.locals.user;
-  res.render('mypage.ejs', { userId: userData.userId });
+// 마이 페이지 띄우기
+router.get('/mypage', verifyAccessToken, async (req, res) => {
+  const { userId } = res.locals.user;
+  const userData = await Users.findOne({ where: { userId } }); // DB에서 해당 userId를 갖고있는 user의 data를 할당.
+
+  res.render('mypage.ejs', {
+    nickname: userData.nickname,
+  });
 });
 
 // 유저 정보 조회 (유저정보 + 게시글 + 댓글)
-router.get('/mypage/:userId', verifyAccessToken, async (req, res) => {
-  const { userId } = req.params;
+router.get('/mypage/userinfo', verifyAccessToken, async (req, res) => {
   const userData = res.locals.user;
-  console.log(userData);
 
   try {
-    if (userId === String(userData.userId)) {
-      // 유저 정보 조회
-      const user = await Users.findOne({
-        attributes: [
-          'nickname',
-          'email',
-          'interest',
-          [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
-        ],
-        where: { userId },
-      });
+    // 유저 정보 조회
+    const user = await Users.findOne({
+      attributes: [
+        'nickname',
+        'email',
+        'interest',
+        [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
+      ],
+      where: { userId: userData.userId },
+    });
 
-      // 유저의 게시글 조회
-      const post = await Posts.findAll({
-        attributes: [
-          'title',
-          'content',
-          [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
-        ],
-        where: { UserId: userId },
-      });
+    // 유저의 게시글 조회
+    const post = await Posts.findAll({
+      attributes: [
+        'title',
+        'content',
+        [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
+      ],
+      where: { UserId: userData.userId },
+    });
 
-      // 유저의 댓글 조회
-      const comment = await Comments.findAll({
-        attributes: [
-          'PostId',
-          'content',
-          [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
-        ],
-        where: { UserId: userId },
-      });
+    // 유저의 댓글 조회
+    const comment = await Comments.findAll({
+      attributes: [
+        'PostId',
+        'content',
+        [Sequelize.fn('left', Sequelize.col('createdAt'), 10), 'date'],
+      ],
+      where: { UserId: userData.userId },
+    });
 
-      // 유저의 정보, 게시글, 댓글 조회
-      return res.status(200).json({
-        user,
-        posts: post,
-        comments: comment,
-      });
-    } else {
-      return res.status(401).json({ errorMessage: '회원 정보에 접근권한이 없습니다.' });
-    }
+    // 유저의 정보, 게시글, 댓글 조회
+    return res.status(200).json({
+      user,
+      posts: post,
+      comments: comment,
+    });
   } catch (err) {
     res.status(400).json({ errorMessage: '유저 정보 조회에 실패하였습니다.' });
   }
@@ -66,15 +65,9 @@ router.get('/mypage/:userId', verifyAccessToken, async (req, res) => {
 // 유저 정보 수정 (nickname,interest,password)
 
 // 유저 닉네임 변경코드
-router.put('/mypage/:userId/nickname', verifyAccessToken, async (req, res) => {
+router.put('/mypage/nickname', verifyAccessToken, async (req, res) => {
   const userData = res.locals.user;
-  const { userId } = req.params;
   const { nickname } = req.body;
-
-  // 닉네임 변경 권한 확인
-  if (userData.userId !== Number(userId)) {
-    return res.status(412).json({ errorMessage: '닉네임 변경권한이 없습니다.' });
-  }
 
   // 닉네임 구성요소 확인
   try {
@@ -87,7 +80,6 @@ router.put('/mypage/:userId/nickname', verifyAccessToken, async (req, res) => {
 
     // 닉네임 중복확인
     const existNickname = await Users.findAll({ where: { nickname } });
-    // console.log(existNickname.length);
     if (existNickname.length !== 0) {
       return res.status(412).json({
         errorMessage: '이미 존재하는 닉네임 입니다.',
@@ -103,27 +95,34 @@ router.put('/mypage/:userId/nickname', verifyAccessToken, async (req, res) => {
 });
 
 // 유저 관심사 변경코드
-router.put('/mypage/:userId/interest', verifyAccessToken, async (req, res) => {
+router.put('/mypage/interest', verifyAccessToken, async (req, res) => {
   const userData = res.locals.user;
-  const { userId } = req.params;
   const { interest } = req.body;
 
+  // 관심사 구성요소
   const interestArr = ['Music', 'Restaurant', 'Exercise', 'Movie', 'Travel'];
+  // 구성요소 파악을 위해 대문자로 변환
+  const upperCaseInterestArr = interestArr.map((i) => i.toUpperCase());
 
   try {
-    // 관심사 변경 권한 확인
-    if (userData.userId !== Number(userId)) {
-      return res.status(412).json({ errorMessage: '관심사 변경권한이 없습니다.' });
-    }
-
     // 관심사 형식 예외 처리
     if (!interest || interest.includes(' ') || interest === '' || interest === undefined) {
       return res.status(412).json({ errorMessage: '작성한 관심사의 형식이 올바르지 않습니다.' });
     }
 
     // 관심사 리스트 외의 요소들 예외 처리
-    if (!interestArr.includes(interest)) {
+    if (!upperCaseInterestArr.includes(interest)) {
       return res.status(412).json({ errorMessage: '작성한 관심사가 리스트에 존재하지 않습니다.' });
+    }
+
+    // 현재 관심사와 변경한 관심사가 일치할 경우
+    // userData.interest에 있는 데이터가 DB에 저장되어 있는 interest 데이터와 달라서, find로 찾아옴.
+    const user = await Users.findAll({
+      where: { userId: userData.userId },
+    });
+    // 예외처리
+    if (user[0].interest === interest) {
+      return res.status(412).json({ errorMessage: '작성하신 관심사가 현재 관심사와 동일합니다.' });
     }
 
     // 관심사 변경
@@ -136,11 +135,9 @@ router.put('/mypage/:userId/interest', verifyAccessToken, async (req, res) => {
 });
 
 // 유저 패스워드 변경코드
-router.put('/mypage/:userId/password', verifyAccessToken, async (req, res) => {
+router.put('/mypage/password', verifyAccessToken, async (req, res) => {
   const userData = res.locals.user;
-  const { userId } = req.params;
   const { newPassword, confirm } = req.body;
-  // console.log();
 
   // 바디에 입력받은 newPassword 해쉬화
   const saltRounds = 10;
@@ -150,7 +147,7 @@ router.put('/mypage/:userId/password', verifyAccessToken, async (req, res) => {
   console.log(hashedPassword);
 
   // Users 모델에 저장되어 있는 userId의 패스워드
-  const { password } = await Users.findOne({ where: { userId } });
+  const { password } = await Users.findOne({ where: { userId: userData.userId } });
   console.log(password);
 
   // 해쉬화 시킨 패스워드와, 로그인된 유저의 패스워드가 일치한지 확인
@@ -158,11 +155,6 @@ router.put('/mypage/:userId/password', verifyAccessToken, async (req, res) => {
   console.log(matchPassword);
 
   try {
-    // 패스워드 변경 권한 확인
-    if (userData.userId !== Number(userId)) {
-      return res.status(412).json({ errorMessage: '패스워드 변경권한이 없습니다.' });
-    }
-
     // 패스워드형식 예외처리
     if (
       !newPassword ||
@@ -210,25 +202,36 @@ router.put('/mypage/:userId/password', verifyAccessToken, async (req, res) => {
 });
 
 // 회원탈퇴
-router.delete('/mypage/:userId', verifyAccessToken, async (req, res) => {
+router.delete('/mypage/userfire', verifyAccessToken, async (req, res) => {
   const userData = res.locals.user;
-  const { userId } = req.params;
-  const { password, confirm } = req.body;
+  const { deletePassword } = req.body;
+
+  // 바디에 입력받은 deletePassword 해쉬화
+  const saltRounds = 10;
+
+  // 바디에 입력받은 deletePassword 단방향 암호화
+  const hashedPassword = await bcrypt.hash(deletePassword, saltRounds);
+  console.log(hashedPassword);
+
+  // Users 모델에 저장되어 있는 userId의 패스워드
+  const { password } = await Users.findOne({ where: { userId: userData.userId } });
+  console.log(password);
+
+  // 해쉬화 시킨 패스워드와, 로그인된 유저의 패스워드가 일치한지 확인
+  const matchPassword = await bcrypt.compare(deletePassword, password);
+  console.log(matchPassword);
 
   try {
-    // 회원탈퇴 권한 확인
-    if (userData.userId !== Number(userId)) {
-      return res.status(412).json({ errorMessage: '회원탈퇴 권한이 없습니다.' });
-    }
-
-    // password가 confirm값과 일치하면 회원탈퇴 진행.
-    if (password !== confirm) {
+    // 입력받은 password가 회원의 패스워드와 일치한지 확인
+    if (!matchPassword) {
       return res.status(412).json({ errorMessage: '패스워드 확인값이 일치하지 않습니다.' });
     }
 
     // 회원탈퇴
-    await Users.destroy({ where: { userId } });
-    return res.status(200).json({ message: '회원탈퇴가 정상적으로 완료되었습니다.' });
+    if (matchPassword) {
+      await Users.destroy({ where: { userId: userData.userId } });
+      return res.status(200).json({ message: '회원탈퇴가 정상적으로 완료되었습니다.' });
+    }
   } catch (err) {
     console.log(err);
     return res.status(400).json({ errorMessage: '회원탈퇴가 정상적으로 처리되지 않았습니다.' });
